+++ NEWS.md
@@ -1,3 +1,6 @@
+# detectMutation - bordeaux data 0.1.9006
+* original mutations positions
+
 # detectMutation - bordeaux data 0.1.9005
 * singularity image for R scripts